requirements.txt
python-telegram-bot==20.6
httpx==0.25.2
aiohttp==3.10.10
